Add utils and expression_base file
