+++ packages/expression_language/lib/src/expressions/expression_base.dart
@@ -1,0 +1,7 @@
+import '../visitors/expression_visitor.dart';
+
+abstract class Expression<T> {
+  T evaluate();
+  void accept(ExpressionVisitor visitor);
+  Type getType() => T;
+}
